fix: reset password page
